refactor scalar lemmas
